+++ faceAILib/src/main/AndroidManifest.xml
@@ -43,13 +43,13 @@
 
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.verify.LivenessDetectActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.verify.LivenessDetectActivity"
             android:theme="@style/FaceAITheme"
             android:exported="false"/>
 
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.verify.FaceVerification32CPUTestActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.verify.FaceVerification32CPUTestActivity"
             android:theme="@style/FaceAITheme"
             android:exported="false"/>
 
@@ -60,7 +60,7 @@
             android:exported="false" />
 
         <activity
-            android:name=".SysCamera.camera.CustomCameraActivityAbs"
+            android:name=".SysCamera.camera.CustomCameraActivity"
             android:theme="@style/AppTheme"
             android:exported="false" />
 
@@ -71,27 +71,25 @@
         </activity>
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.verify.FaceVerificationActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.verify.FaceVerificationActivity"
             android:theme="@style/FaceAITheme"
             android:exported="false"/>
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.addFace.AddFaceImageActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.addFace.AddFaceImageActivity"
             android:theme="@style/FaceAITheme"
             android:exported="false" />
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.search.FaceSearch1NActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.search.FaceSearch1NActivity"
             android:theme="@style/FaceAITheme"
             android:exported="false" />
 
         <activity
-            android:name="com.faceAI.demo.SysCamera.search.FaceSearchMNActivityAbs"
+            android:name="com.faceAI.demo.SysCamera.search.FaceSearchMNActivity"
             android:exported="false"
             android:theme="@style/AppTheme"
             android:launchMode="singleTop" />
-
-
 
         <activity
             android:name="com.faceAI.demo.SysCamera.verify.FaceVerifyWelcomeActivity"
