+++ faceAILib/src/main/java/com/faceAI/demo/SysCamera/verify/FaceVerifyWelcomeActivity.java
@@ -1,8 +1,8 @@
 package com.faceAI.demo.SysCamera.verify;
 
 import static com.faceAI.demo.FaceImageConfig.CACHE_BASE_FACE_DIR;
-import static com.faceAI.demo.SysCamera.addFace.AddFaceImageActivityAbs.ADD_FACE_IMAGE_TYPE_KEY;
-import static com.faceAI.demo.SysCamera.verify.FaceVerificationActivityAbs.USER_FACE_ID_KEY;
+import static com.faceAI.demo.SysCamera.addFace.AddFaceImageActivity.ADD_FACE_IMAGE_TYPE_KEY;
+import static com.faceAI.demo.SysCamera.verify.FaceVerificationActivity.USER_FACE_ID_KEY;
 
 import android.app.AlertDialog;
 import android.content.Intent;
@@ -24,7 +24,7 @@ import com.chad.library.adapter.base.listener.OnItemLongClickListener;
 import com.faceAI.demo.UVCCamera.verify.FaceVerify_UVCCameraActivity;
 import com.faceAI.demo.UVCCamera.addFace.AddFace_UVCCameraActivity;
 import com.faceAI.demo.UVCCamera.addFace.AddFace_UVCCameraFragment;
-import com.faceAI.demo.SysCamera.addFace.AddFaceImageActivityAbs;
+import com.faceAI.demo.SysCamera.addFace.AddFaceImageActivity;
 import com.faceAI.demo.SysCamera.search.ImageBean;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
@@ -79,8 +79,8 @@ public class FaceVerifyWelcomeActivity extends AbsAddFaceFromAlbumActivity {
         addFaceView.setOnClickListener(view -> {
                     if (dataSourceType.equals(DataSourceType.Android_HAL)) {
                         startActivity(
-                                new Intent(getBaseContext(), AddFaceImageActivityAbs.class)
-                                        .putExtra(ADD_FACE_IMAGE_TYPE_KEY, AddFaceImageActivityAbs.AddFaceImageTypeEnum.FACE_VERIFY.name()));
+                                new Intent(getBaseContext(), AddFaceImageActivity.class)
+                                        .putExtra(ADD_FACE_IMAGE_TYPE_KEY, AddFaceImageActivity.AddFaceImageTypeEnum.FACE_VERIFY.name()));
                     } else {
                         startActivity(
                                 new Intent(getBaseContext(), AddFace_UVCCameraActivity.class)
@@ -120,7 +120,7 @@ public class FaceVerifyWelcomeActivity extends AbsAddFaceFromAlbumActivity {
             @Override
             public boolean onItemLongClick(@NonNull BaseQuickAdapter<?, ?> adapter, @NonNull View view, int i) {
                 startActivity(
-                        new Intent(getBaseContext(), FaceVerification32CPUTestActivityAbs.class)
+                        new Intent(getBaseContext(), FaceVerification32CPUTestActivity.class)
                                 .putExtra(USER_FACE_ID_KEY, faceImageList.get(i).name));
                 return false;
             }
@@ -131,7 +131,7 @@ public class FaceVerifyWelcomeActivity extends AbsAddFaceFromAlbumActivity {
                     // 根据摄像头种类启动不同的模式
                     if (dataSourceType.equals(DataSourceType.Android_HAL)) {
                         startActivity(
-                                new Intent(getBaseContext(), FaceVerificationActivityAbs.class)
+                                new Intent(getBaseContext(), FaceVerificationActivity.class)
                                         .putExtra(USER_FACE_ID_KEY, faceImageList.get(i).name));
                     } else {
                         //USB UVC协议摄像头，双目
