+++ faceAILib/src/main/res/layout/activity_face_search.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:customAttr="http://schemas.android.com/apk/res-auto"
-    tools:context=".SysCamera.search.FaceSearch1NActivityAbs">
+    tools:context=".SysCamera.search.FaceSearch1NActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_camerax"
